league booking 추가

+++ kick-off-batch/src/main/java/com/kickoff/batch/config/feign/api/temp/Bookings.java
@@ -1,9 +1,7 @@
 package com.kickoff.batch.config.feign.api.temp;
 
-import com.kickoff.batch.config.feign.api.MatchesResultDetailResponse;
-
 public record Bookings(
-        int minute,
+        long minute,
         DetailTeam team,
         Player player,
         String card
